Add files via upload
Case Study: Ride Sharing Analysis
This data set is a masked and simplified data set which is like what data analysts at Uber/Lyft handle. 
Business Understanding
You may have some experience of traveling to and from the airport. Have you ever used Uber/Lyft or any other cab service for this trip? Did you ‌face the problem of cancellation by the driver or non-availability of cars?
These problems also affect the ride companies, as their revenue is tied to the trips made. 
We will investigate this problem of non availability and cancellation 
Business Objectives
The aim of analysis is to identify the root cause of the problem (i.e. cancellation and non-availability of cars) and recommend ways to improve the situation. 
As a result of your analysis, you should be able to highlight the root cause(s) and possible hypotheses of the problem(s) and recommend ways to improve them.
Data download: 
https://raw.githubusercontent.com/vkoul/data/main/misc/ride_data.csv
Data Dictionary:
There are six attributes associated with each request made by a customer:
Request id: A unique identifier of the request 
Pick-up point: The point from which the request was made 
Driver id: The unique identification number of the driver 
Status: The final status of the trip, that can be either completed, canceled by the driver or no cars available
Request Time Stamp: The date and time at which the customer made the trip request 
Drop Time Stamp: The drop-off date and time, in case the trip was completed